refactor: OAuth2 config 관련 수정
