+++ sellyourunhappiness-api/src/main/java/sellyourunhappiness/global/config/security/oauth2/CustomOAuth2User.java
@@ -1,4 +1,4 @@
-package sellyourunhappiness.global.config.security.oauth;
+package sellyourunhappiness.global.config.security.oauth2;
 
 import java.util.Collection;
 import java.util.Map;
